requirements.txt
# requirements.txt
# Install all dependencies with:
#   pip install -r requirements.txt

requests==2.32.3        # Making HTTP API calls
pandas==2.2.3           # Working with tabular data
psycopg[binary]>=3.1    # Connecting to Redshift (PostgreSQL-compatible driver, supports Python 3.14)
python-dotenv==1.0.1    # Loading credentials from the .env file
